perf(itun): add route-level code splitting with lazy loading
Split 3 large routes into config + lazy component files using
TanStack Router's createLazyFileRoute:
- Game detail page (games/$gameId/index)
- Create crawler wizard (games/$gameId/create-crawler)
- Create mech wizard (pilots/$pilotId/create-mech)

Each route config now has pendingComponent: PageSkeleton for loading
states during route transitions. 19 new tests verify file structure.
